requirements/dev.txt
-r base.txt

pytest-django==3.7.0
prospector==1.1.7
factory_boy==2.12.0
Faker==2.0.3
django-debug-toolbar==2.0
pylint==2.4.3
black==19.10b0
pre-commit==1.20.0
fakeredis==1.0.5
